chore: rename primary_screening package to screening
